pyproject.toml
[project]
name = "scrappersally"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.57.0",
    "streamlit>=1.52.1",
    "yt-dlp>=2025.12.8",
    "python-dotenv>=1.0.1",
]
